requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
Django>=5.1,<6.0
djangorestframework>=3.15
django-environ>=0.11
gunicorn>=23.0
whitenoise>=6.8
psycopg2-binary>=2.9

# ── Async / Task Queue ────────────────────────────────────────────────────────
celery>=5.4
redis>=5.0

# ── File Storage (Cloudflare R2 via S3-compatible API) ───────────────────────
django-storages[s3]>=1.14
boto3>=1.35

# ── Auth ──────────────────────────────────────────────────────────────────────
djangorestframework-simplejwt>=5.3

# ── API Docs ──────────────────────────────────────────────────────────────────
drf-spectacular>=0.27

# ── Transcription ─────────────────────────────────────────────────────────────
assemblyai>=0.33

# ── PII Redaction ─────────────────────────────────────────────────────────────
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
spacy>=3.7

# ── AI SOAP Generation ────────────────────────────────────────────────────────
groq>=0.13
pydantic>=2.0

# ── PDF Export ────────────────────────────────────────────────────────────────
weasyprint>=62

# ── Utilities ─────────────────────────────────────────────────────────────────
Pillow>=10.0
requests>=2.32
